% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoacids_df.R
\docType{data}
\name{aminoacids_df}
\alias{aminoacids_df}
\title{Amino acid abbreviations}
\format{A data frame with 27 rows and 4 columns:
\describe{
 \item{one}{One-letter codes of amino acids}
 \item{three}{Three-letter codes of amino acids}
 \item{full}{Full name of the amino acid}
 \item{amb}{Logical indicating if abbreviation is ambiguous, i.e., matches 
 more than one amino acid}
 }}
\source{
Johnson, A.D. (2010). An extended IUPAC nomenclature code for 
polymorphic nucleic acids. Bioinformatics 26, 1386–1389.
}
\description{
A dataset containing the one- and three-letter codes, and full names of 
amino acids according to the IUPAC nomenclature.
}
\examples{
data(aminoacids_df)

}
\keyword{datasets}
