\name{tidysq-package}
\alias{tidysq-package}
\alias{tidysq}
\docType{package}
\title{\packageTitle{tidysq}}
\description{\packageDescription{tidysq}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{tidysq}
  \packageIndices{tidysq}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{tidysq}
  
  Maintainer: \packageMaintainer{tidysq}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
