% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_methods.R
\name{as.character.sq}
\alias{as.character.sq}
\title{Convert sq object into character vector}
\usage{
\method{as.character}{sq}(x, ...)
}
\arguments{
\item{x}{a \code{\link{sq}} object to be converted}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
A \code{character} vector of the length the same as number
of sequences in the converted \code{sq} object
}
\description{
Coerce sequences from a \code{\link{sq}} object to \code{\link{character}} vector
of sequences
}
\details{
This method for class \code{\link{sq}} allows converting sequences from
the sq object into a character vector of length equal to the length 
of sq. Each element of resulting vector is a separate sequence. 
All attributes of the input sq are lost during the conversion to 
character vector.
}
\examples{
# Creating an object to work on:
sq_nuc <- construct_sq(c("CTGAATGCAGTACCGTAAT", "ATGCCGTAAATGCCAT", 
                         "CAGACCANNNATAG"), type = 'nuc')
                         
# Converting the sq object into a character vector:
as.character(sq_nuc)

}
\seealso{
sq
}
