% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-methods.R
\name{as.sq}
\alias{as.sq}
\title{Convert an object to sq}
\usage{
as.sq(x, ...)
}
\arguments{
\item{x}{[\code{any}]
An object of a class that supports conversion to \code{\link[=sq-class]{sq class}}.}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
A \code{sq} object
}
\description{
This generic function takes an object of arbitrary type and returns an \code{\link[=sq-class]{sq}} object
as an output. Default implementation of the method throws an error - there needs to be 
implemented a method for specified class in order for function to work.
}
\details{
In \pkg{tidysq} only for \code{\link{character}} vector there is an implemented method 
which in fact works exactly like \code{\link{construct_sq}} - you can also pass other arguments
like those supported by \code{\link{construct_sq}}.
}
\seealso{
\code{\link[=sq-class]{sq}} \code{\link{construct_sq}}
}
