% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUM50_enc}
\alias{BLOSUM50_enc}
\title{The BLOSUM50_enc matrix}
\format{A data frame with with 21 rows and 21 columns.
\describe{Contains an one-letter codes of amino acid as columns and rows names
and also row and column zero vectors representing 'X' (any amino acid).
You can check three-letter codes and full names of the amino acids in
\code{\link{aminoacids_df}}.
 }}
\source{
Henikoff, Steven, and Jorja G. Henikoff. 
"Amino acid substitution matrices from protein blocks." 
Proceedings of the National Academy of Sciences 89.22 (1992): 10915-10919.
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM50}
}
\description{
The BLOck SUbstition Matrix (BLOSUM) for sequences with less than 50\% similarity
transformed for encoding by dividing each value by 5.
The matrix has been subset to the 20 proteogenic amino acids.
}
\details{
BLOSUM matrix is a substitution matrix used for sequence alignment of proteins.
BLOSUM matrices are actual percentage identity values of sequences selected for
construction of the matrices. BLOSUM50 indicates that the sequences
selected for constructing the matrix share an average identity value of 50\%.
BLOSUM50 is good matrix for distantly related proteins.
The matrix made by matblas from blosum50.iij
BLOSUM Clustered Scoring Matrix in 1/3 Bit Units
Cluster Percentage: >= 50
Entropy =   0.4808, Expected =  -0.3573
}
\examples{
data(BLOSUM50_enc)

}
\seealso{
aminoacids_df Blosum50
}
\keyword{datasets}
\keyword{matrix}
