% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLOSUM_data.R, R/data.R
\docType{data}
\name{BLOSUM50_pca}
\alias{BLOSUM50_pca}
\title{The BLOSUM50_pca matrix}
\format{A data frame with with 21 rows and 20 columns.
\describe{Contains an one-letter codes of amino acid as rows (additionaly 'X' as any amino acid)
and ordinal numbers of principal component direction as columns names. Twenty columns represent 
the loadings of the twenty eigenvectors. 
You can check three-letter codes and full names of the amino acids in
\code{\link{aminoacids_df}}.
 }}
\source{
Li, Jie & Koehl, Patrice. (2014). 3D representations of amino acids - 
Applications to protein sequence comparison and classification. 
Computational and Structural Biotechnology Journal. 11. 10.1016/j.csbj.2014.09.001. 
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM50}

Li, Jie & Koehl, Patrice. (2014). 3D representations of amino acids - 
Applications to protein sequence comparison and classification. 
Computational and Structural Biotechnology Journal. 11. 10.1016/j.csbj.2014.09.001. 
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM50}
}
\description{
This matrix enables principal component analysis (PCA) with usage \code{\link{BLOSUM50}}.

This matrix enables principal component analysis (PCA) with usage \code{\link{BLOSUM50}}.
}
\details{
The BLOSUM50_pca matrix enables PCA calculation on proteins sequences aligments.
Components are generated by an eigenvector decomposition of the matrix formed
from pairwise similarity scores between each pair of sequences. The similarity score model
used for creating BLOSUM50_pca matrix is the \code{\link{BLOSUM50}}.

The BLOSUM50_pca matrix enables PCA calculation on proteins sequences aligments.
Components are generated by an eigenvector decomposition of the matrix formed
from pairwise similarity scores between each pair of sequences. The similarity score model
used for creating BLOSUM50_pca matrix is the \code{\link{BLOSUM50}}.
}
\examples{
data(BLOSUM50_pca)

data(BLOSUM50_pca)

}
\seealso{
aminoacids_df Blosum50

aminoacids_df Blosum50
}
\keyword{datasets}
\keyword{matrix}
