% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{BLOSUM62_enc}
\alias{BLOSUM62_enc}
\title{The BLOSUM62_enc matrix}
\format{A data frame with with 21 rows and 21 columns.
\describe{Contains an one-letter codes of amino acid as columns and rows names
and also row and column zero vectors representing 'X' (any amino acid).
You can check three-letter codes and full names of the amino acids in
\code{\link{aminoacids_df}}.
 }}
\source{
Henikoff, Steven, and Jorja G. Henikoff. 
"Amino acid substitution matrices from protein blocks." 
Proceedings of the National Academy of Sciences 89.22 (1992): 10915-10919.
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM62}
}
\description{
The BLOck SUbstition Matrix (BLOSUM) for sequences with less than 62\% similarity.
The matrix has been subset to the 20 proteogenic amino acids.
}
\details{
BLOSUM matrix is a substitution matrix used for sequence alignment of proteins.
BLOSUM matrices are actual percentage identity values of sequences selected for
construction of the matrices. BLOSUM62 indicates that the sequences
selected for constructing the matrix share an average identity value of 62\%.
BLOSUM62 is miderange matrix between close and distantly related proteins.
Matrix made by matblas from blosum62.iij
BLOSUM Clustered Scoring Matrix in 1/2 Bit Units
Cluster Percentage: >= 62
Entropy = 0.6979, Expected = -0.5209
}
\examples{
data(BLOSUM62_enc)

}
\seealso{
aminoacids_df BLOSUM62
}
\keyword{datasets}
\keyword{matrix}
