% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_methods.R
\name{c.sq}
\alias{c.sq}
\title{Concatenate sq objects}
\usage{
\method{c}{sq}(...)
}
\arguments{
\item{...}{multiple \code{\link{sq}} objects. All of them have to have the same type and
subtype. If type is \strong{atp}, \strong{unt} or \strong{enc} also their alphabets have to
be exactly identical.}
}
\value{
A \code{\link{sq}} object with length equal to sum of lengths of individual objects
passed as parameters. Elements of \code{\link{sq}} are concatenated just as if they were normal
lists (see \code{\link[base]{c}})
}
\description{
Merges multiple \code{\link{sq}} objects of the same type into one larger.
}
\examples{
sq_1 <- construct_sq(c("TAGACTAG", "", "CCGTAGATG"))
sq_2 <- construct_sq(c("TTGATAACG", "TGTATGTGA"))
sq_3 <- construct_sq(character(0))
sq_4 <- construct_sq("gaGG")

c(sq_1, sq_2, sq_3, sq_4)

}
