% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean sequences containing ambiguous elements}
\usage{
clean(sq, only_elements = FALSE)
}
\arguments{
\item{sq}{\code{\link{sq}} object of type 'ami' or 'nuc'}

\item{only_elements}{logical indicating if only ambiguous elements
(i.e., matching more than one amino acid/nucleotide) of sequences should
be removed. If \code{FALSE} (default) whole sequences containing ambiguous 
elements are removed.}
}
\value{
\code{\link{clnsq}} object of the same type as input sq.
}
\description{
Remove sequences containing ambiguous elements or remove ambiguous 
elements from sequences in sq object.
}
\details{
This function allows cleaning of sequences containing ambiguous
elements. By default, sequences containing ambiguous elements are removed 
and \code{NULL sq} values are introduced in their place. If \code{only_elements = TRUE} 
then only ambiguous elements are removed from sequences in sq object.
}
\examples{
# creating objects to work on:
sq_ami <- construct_sq(c("MIAANYTWIL","TIAALGNIIYRAIE", "NYERTGHLI", "MAYXXXIALN"), type = "ami")
sq_nuc <- construct_sq(c("ATGCAGGA", "GACCGAACGAN", "TGACGAGCTTA", "ACTNNAGCN"), type = "nuc")

# removing sequences with ambiguous elements:
clean(sq_ami)
clean(sq_nuc)

# removing ambiguous elements from sequences:
clean(sq_ami, only_elements = TRUE)
clean(sq_nuc, only_elements = TRUE)

}
\seealso{
sq clnsq
}
