% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest}
\alias{digest}
\title{Digest proteins at specified cleavage sites}
\usage{
digest(ami_sq, name, digest_pattern)
}
\arguments{
\item{ami_sq}{a \code{\link{sq}} object with \strong{ami} type and \strong{cln} subtype.}

\item{name}{a non-\code{NULL} \code{character} vector without \code{\link{NA}} values, 
containing names of the sequences in the sq. It has to be of the same length 
as the \code{sq}.}

\item{digest_pattern}{a \code{character} protease cleavage site pattern.}
}
\value{
A \code{\link[tibble]{tibble}} with following columns:
 \item{name}{name of the sequence}
 \item{peptide}{fragment of the sequence, result of digesting}
}
\description{
TEMPORARY DOCS
}
\details{
TEMPORARY DOCS
}
