% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm.R
\name{empty_pssm}
\alias{empty_pssm}
\title{Create empty Position Specific Scoring Matrix}
\usage{
empty_pssm(num_pos, type)
}
\arguments{
\item{num_pos}{the number of positions in a sequence.}

\item{type}{\code{character} type of a sequence}
}
\value{
A numeric matrix with rows equal to \code{num_pos}
and number of columns equal to the number of characters in the specified 
alphabet.
}
\description{
Create an empty Position Specific Scoring Matrix 
for a specified alphabet.
}
\examples{
# empty nucleotide PSSM
empty_pssm(num_pos = 9, type = "nuc")

# empty aminoacid PSSM
empty_pssm(num_pos = 9, type = "ami")
}
