% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_methods.R
\name{==.sq}
\alias{==.sq}
\title{Compare sq objects}
\usage{
\method{==}{sq}(x1, x2)
}
\arguments{
\item{x1}{an \code{\link{sq}} object.}

\item{x2}{an object (a character vector or an \code{\link{sq}} object) to compare with
\code{x1}.}
}
\value{
A \code{\link{logical}} vector indicating on which positions the objects are equal
}
\description{
Compares input \code{\link{sq}} object with another given.
}
\details{
\code{`==`} converts left-hand side of the comparison (\code{x1}) to
a character vector using \code{\link{as.character}} and checks whether object
given on the right side can be compared with \code{\link{sq}} object. Function
also checks the type of \code{\link{sq}} object with which given object is to be
compared. If the type of \code{\link{sq}} object is "ami", "dna" or "rna" and
given sequence is a character vector with any lowercase letters, the function
replaces them with corresponding capital ones using \code{\link{toupper}}.
If right-hand side object (\code{x2}) is of class \code{\link{sq}}, then it is
converted to a character vector also using \code{\link{as.character}} function.

When both objects are already converted to character vectors, comparison is carried out 
element-wise with standard R rules, (e.g., recycling is used). You can check details at
\code{\link{Comparison}}.

Comparing sequences as characters vectors cause that various types of sequences
can be compared for example amino acid with nucleotide sequences so attention
should be paid, which sequence types are compared.
}
\examples{

# Creating sq object to work on:
sq <- construct_sq(c("ACTGCTG", "CTTAGA", 
                     "CCCT", "CTGAATGT"), type = "dna")
                     
sq_different_len <- construct_sq(c("ACTGCTG", "CTTAGA", 
                                   "GGAA"), type = "dna")
                                   
sq_the_same_len <- construct_sq(c("ACTGCTG", "CTTAGA", 
                                 "CCCT", "CTGAATGT"), type = "dna")
                                                                                       
# Get an overview of the sequences:
summary(sq)
summary(sq_the_same_len)
summary(sq_different_len)

# Comparing sq object with an object of the same length :
sq == sq_the_same_len

# Comparing object sq object with an object of a different length : 
sq == sq_different_len
 
# Comparing sq object to a given character vector of a different length:
sq == c('AAA','CCC')

# Comparing sq object to a given character vector of the same length:
sq == c("ACTGCTG", "CTTAGA",'CCCT', 'CTGAATGT')

# Comparing sq object to a given DNA element 'ATGTGA':
sq == 'ATGTGA'

# Comparing sq object to a given amino acid vector:
sq == c('RISGQQD','RISGQQD')
 
}
\seealso{
\code{\link{sq}} \code{\link{as.character}} \code{\link{is.sq}}
}
