% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{extract_kmers}
\alias{extract_kmers}
\title{Extract k-mers from sequences}
\usage{
extract_kmers(sq, lens = 0, dists = list(0))
}
\arguments{
\item{sq}{a vector or matrix describing sequence(s).}

\item{lens}{\code{integer} vector of lengths of k-mers.}

\item{dists}{\code{integer} vector of distances between elements of k-mer (0 means 
consecutive elements). See Details.}
}
\value{
A \code{character} matrix of k-mers, where every row corresponds to a
different sequence.
}
\description{
Extracts vector of k-mers present in sequence(s).
}
\details{
A format of \code{d} vector is discussed in Details of 
\code{\link{count_kmers}}.
}
\examples{
set.seed(1410)
random_sqs <- random_sq(3, 5, type = "nuc", is_clean = TRUE)

# 1-mers
extract_kmers(random_sqs, lens = 1, dists = list(0))

# 2-mers with and without gap
extract_kmers(random_sqs, lens = c(2, 2), dists = list(0, 1))
}
