% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tidysq_options.R
\name{fast-mode}
\alias{fast-mode}
\title{Make your operations faster}
\description{
\emph{Fast mode} is meant to improve performance of operations in package. However, turning it 
on is associated with less control - some of parameters checks are skipped, so user has to be
certain that they are correct.
}
\details{
Not all functions support \emph{fast mode} yet. Those that do not always operate in normal mode.
}
\section{Turning fast mode on and off}{


\code{Fast mode} can be turned on and off by setting package option "tidysq_fast_mode" to 
accordingly \code{TRUE} and \code{FALSE} with the latter value as default. To learn more about
options, see \code{\link{tidysq-options}}.
}

\section{Functions supporting fast mode}{


Below there is a list of functions that do support \emph{fast mode}:
\itemize{
\item \code{\link{construct_sq}}
\item \code{\link{read_fasta}}
}
}

\seealso{
\code{\link{sq}} \code{\link{tidysq-options}}
}
