% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_invalid_letters.R
\name{find_invalid_letters}
\alias{find_invalid_letters}
\alias{find_invalid_letters.sq}
\title{Find elements which are not suitable for specified type.}
\usage{
find_invalid_letters(x, dest_type, ...)

\method{find_invalid_letters}{sq}(
  x,
  dest_type,
  ...,
  NA_letter = getOption("tidysq_NA_letter")
)
}
\arguments{
\item{x}{a \code{\link[=sq-class]{sq}} object to be checked.}

\item{dest_type}{a \code{\link{character}} string denoting destination type - it may be 
"dna" for \strong{dna} type (DNA), "rna" for \strong{rna} type (RNA) or
"ami" for \strong{ami} type (amino acids).}
}
\value{
A list of mismatched elements for every sequence from \code{\link[=sq-class]{sq}} object.
}
\description{
Function finds elements in given sequence not matching to amino acid or nucleotide 
alphabet.
}
\details{
This function allows obtaining list of mismatched elements of sequences from  
a \code{\link[=sq-class]{sq}} object to amino acid, DNA or RNA alphabet. Output list has number of
elements equal to length of \code{sq} object and each element is a character vector 
of elements that appear in according sequence that does not fit destination type. This 
function might be used to find specifically, which sequences have letters - user
may want to use this information for example to check input sequences.

You can check, which letters are valid for specified type in \code{\link[=sq-class]{sq}} class
documentation.

Mismatched elements might be replaced with other letters or \code{NA} using 
\code{\link{substitute_letters}} and then, after removal of all non-standard letters,
set type to destinated using \code{\link{typify}}.

Returned lists for \strong{ami}, \strong{dna} and \strong{rna} \code{sq} objects,
if \code{des_type} is specified respectively "ami", "dna" and "rna", will contain
only \code{\link[=sq-class]{NULL}} elements.
}
\seealso{
\code{\link[=sq-class]{sq}} \code{\link{construct_sq}}
}
