% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm.R
\name{flatten_pssm}
\alias{flatten_pssm}
\title{'Flatten' PSSM to a single row}
\usage{
flatten_pssm(pssm)
}
\arguments{
\item{pssm}{a position specific scoring matrix}
}
\value{
A numeric matrix with 1 row and m * n columns
}
\description{
'Flatten' PSSM to a single row, such that e.g. m x n = 9 x 20 becomes 1 x 180
}
\examples{
flatten_pssm(empty_pssm(num_pos = 9, type = "nuc"))
}
