% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_funs.R
\name{get_invalid_letters}
\alias{get_invalid_letters}
\title{Finds elements which are not present in nucleotide or amino acid alphabet.}
\usage{
get_invalid_letters(sq, dest_type)
}
\arguments{
\item{sq}{\code{\link{sq}} object to be checked.}

\item{dest_type}{indices specifying alphabet. Its may be 
\code{nuc} for nucleotide alphabet or \code{ami} for amino acid alphabet.}
}
\value{
a list of mismatched elements for every sequence from \code{\link{sq}} object.
}
\description{
Function finds elements in given sequence not matching to amino acid or nucleotide 
alphabet.
}
\details{
This function allows returning list of mismatched elements from  
\code{\link{sq}} object to amino acid or nucleotide alphabet. Function creates list
for each sequence from \code{\link{sq}} object and append to each list  element not 
matching to any alphabet.
}
\examples{
# Creating an object to work on:       
sq_nucleotides <- construct_sq(c("ACGPOIUATTAGACG","GGATFGHA"), type = "nuc")
sq_amino_acids <- construct_sq(c("QWERTYUIZXCVBNM","LKJHGFDSAZXCVBN"), type = "ami")

# Creating lists of mismatched elements from nucleotide sq object:
get_invalid_letters(sq_nucleotides)

# Creating lists of mismatched elements from amino acid sq object:
get_invalid_letters(sq_amino_acids)

}
\seealso{
sq construct_sq
}
