% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_funs.R
\name{get_sq_alphabet}
\alias{get_sq_alphabet}
\title{Returns full alphabet of type of given sequence.}
\usage{
get_sq_alphabet(sq)
}
\arguments{
\item{sq}{\code{\link{sq}} object to be recognized.}
}
\value{
a character vector of amino acid or nucleotide alphabet elements.
}
\description{
Function returns amino acid or nucleotide alphabet based on sequence type.
}
\details{
This function allows returning character vector of amino acid or 
nucleotide alphabet. The function read provided \code{\link{sq}} object and determines
which kind of sequence user assigned to a \code{\link{sq}} object (nucleotide or amino acid).
If  \code{\link{sq}} contains an amino acid sequence the function returns set of 20 aminoacids with
gap (-) and any amino acid (*) element. If a \code{\link{sq}} contains
a nucleotide sequence the function returns set of 5 nucleotides with gap (-) element.
If \code{\link{sq}} type is defined as "unt" function returns list of elements present
in \code{\link{sq}} object. The details about amino acid and nucleotide alphabet can be
checked in \code{\link{aminoacids_df}} and \code{\link{nucleotides_df}} respectively.
}
\examples{
# Creating an object to work on:
sq_nucleotides <- construct_sq(c("ACGATTAGACG","GGATA"), type = "nuc")
sq_amino_acids <- construct_sq(c("MMVTAAV"), type = "ami")
sq_without_type <- construct_sq(c("ACGATTAGACG"))

# Testing nucleotide sq object with defined type:
get_sq_alphabet(sq_nucleotides)

# Testing amino acid sq object with defined type:
get_sq_alphabet(sq_amino_acids)

# Testing nucleotide sq object without defined type:
get_sq_alphabet(sq_without_type)

  
}
\seealso{
sq construct_sq
}
