% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidysq_options.R
\name{get_tidysq_options}
\alias{get_tidysq_options}
\title{Obtain current state of tidysq options}
\usage{
get_tidysq_options()
}
\value{
a \code{\link[=namedList-class]{named list}} containing values of 
currently set options associated with the \pkg{tidysq} package.
}
\description{
Obtain current state of tidysq options
}
\details{
To see full list of options and their meanings, 
see \code{\link{tidysq-options}} man page.
}
\seealso{
\code{\link{tidysq-options}}
}
