% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sq.R
\name{import_sq}
\alias{import_sq}
\title{Import sq objects from other objects}
\usage{
import_sq(object, ...)
}
\arguments{
\item{object}{- an object of one of classes: \code{AAbin}, \code{DNAbin}, \code{AAStringSet}, 
\code{DNAStringSet}, \code{SeqFastaAA}, \code{SeqFastadna}.}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
A \code{\link[tibble]{tibble}} with \code{sq} column of \code{\link[=sq-class]{sq}} type
representing the same 
sequences as given object; the object has a type corresponding to the input type; if given
sequences had names, output \code{\link[tibble]{tibble}} has also another column 
\code{name} with those names
}
\description{
Creates \code{\link[=sq-class]{sq object}} from \code{object} of class from another package.
Currently supported packages are \pkg{ape} with its formats (\code{AAbin} and \code{DNAbin}),
\pkg{Bioconductor} (\code{AAStringSet}, \code{DNAStringSet}) and
\pkg{seqinr} (\code{SeqFastaAA}, \code{SeqFastadna}).
}
\details{
Providing object of class other than specified will result in error.
}
\seealso{
\code{\link{export_sq}} \code{\link[=sq-class]{sq}}
}
