% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_methods.R
\name{lengths.sq}
\alias{lengths.sq}
\title{Get lengths of sequences in sq object}
\usage{
lengths.sq(x, use.names = TRUE)
}
\arguments{
\item{sq}{a \code{\link{sq}} object.}
}
\value{
A \code{\link{numeric}} vector, where each element gives length of according 
sequence from \code{\link{sq}} object.
}
\description{
Function counts number of elements in each sequence in given \code{\link{sq}} object.
}
\details{
This function allows returning numeric vector of lengths of sequences from
\code{\link{sq}} object. The numeric vector is as long as number of sequences present 
in \code{\link{sq}} object.
The function counts elements in all types of sequences.
}
\examples{
# Creating an object to work on:
sq_dna <- construct_sq(c("ACGATTAGACG","GGATA"), type = "dna")
sq_amino_acids <- construct_sq(c("MMVTAAV"), type = "ami")

# Counting number of elements in DNA sq object with defined type:
lengths(sq_dna)

# Counting number of elements in amino acid sq object with defined type:
lengths(sq_amino_acids)

}
\seealso{
\code{\link{sq}} \code{\link{construct_sq}}
}
