% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nc_pack_cnuc}
\alias{nc_pack_cnuc}
\title{Pack raw bytes, but better}
\usage{
nc_pack_cnuc(UNPACKED)
}
\arguments{
\item{UNPACKED}{\code{raw} vector}

\item{ALPH_SIZE}{\code{integer}}
}
\description{
Pack raw bytes, but better
}
