% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotides_df.R
\docType{data}
\name{nucleotides_df}
\alias{nucleotides_df}
\title{Nucleotides abbreviations}
\format{A data frame with 17 rows and 3 columns:
\describe{
 \item{one}{One-letter codes of nucleotides}
 \item{item}{Full name of the nucleotide}
 \item{amb}{Logical indicating if abbreviation is ambiguous, i.e., matches more than one nucleotide}
 }
 @source IUPAC 
 @keywords datasets
 @examples 
 data(nucleotides_df)}
\description{
A dataset containing the nucleotide letter code and full names od nucleotides according to IUPAC nomenclature
}
