% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pack3}
\alias{pack3}
\title{Pack raw bytes, but better}
\usage{
pack3(UNPACKED, ALPH_SIZE)
}
\arguments{
\item{UNPACKED}{\code{raw} vector}

\item{ALPH_SIZE}{\code{integer}}
}
\description{
Pack raw bytes, but better
}
