% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sq.R
\name{print.sq}
\alias{print.sq}
\title{Print sq object}
\usage{
\method{print}{sq}(x, ...)
}
\arguments{
\item{x}{\code{\link{sq}} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints input \code{\link{sq}} object in a human-friendly form.
}
\details{
\code{Print} checks if the input \code{\link{sq}} object is cleaned and includes this information alongside with type in the printed message. 
This method is used by default in each case of calling the \code{\link{sq}} object.

All \code{\link{NA}} values are replaced with '*' and empty sequences are distinguished.

This is overloaded function from base package. It is selected when \code{\link{sq}} object is used as a parameter for print function. To see the generic function page, 
check \link[here]{https://www.rdocumentation.org/packages/base/versions/3.6.1/topics/print}.
}
\examples{

# Creating sq object to work on:
sq <- construct_sq(c("fafasfasfFSA", "ygagayagfa", "adsDaf"), type = "ami")

# Printing without implicit function calling:
sq

# Printing with implicit function calling:
print(sq)

# Implicit printing of the uncleaned object:
print(construct_sq("ACTAGAGTGATAGTAGGAGTAGA", type = "nuc"))

# Implicit printing of the cleaned object:
print(clean(construct_sq("ACTAGAGTGATAGTAGGAGTAGA", type = "nuc")))

# Implicit printing of the object without defined type:
print(construct_sq(c("afsfd", "q243faadfa", "afsw34gesfv", "adfq2", "fasfas", "g'qp9u2r3'b;")))

# Implicit printing of the object with empty sequence:
print(construct_sq(c("afsfd", "", "adfq2", "fasfas", "")))

# Implicit printing of the object with NA element:
print(construct_sq(c("afsfd", NA, "adfq2", NA, "")))

# Implicit printing of the simplified object:
enc <- c(A = "a", B = "a", C = "a", D = "a", E = "a", F = "b", G = "b", 
         H = "b", I = "c", J = "c", K = "c", L = "c", M = "c", N = "c", 
         O = "c", P = "d", Q = "d", R = "d", S = "d", T = "d", U = "d", 
V = "d", W = "d", X = "d", Y = "d", Z = "d", `-` = "d")
print(simplify(sq, enc))
 
}
\seealso{
\link{sq} \link{clean} \link{simplify}
}
