% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sq.R
\name{random_sq}
\alias{random_sq}
\title{Generate random sequences}
\usage{
random_sq(n, len, type, is_clean, sd = NULL, use_gap = FALSE)
}
\arguments{
\item{n}{a positive \code{\link{integer}} value - number of sequences to generate}

\item{len}{a positive \code{\link{integer}} value - length of each sequence if \code{sd} not 
specified and mean length of sequences if \code{sd} spedified}

\item{type}{type of generated sq object; possible values are "ami" and "nuc" (see section
\emph{sq types} in \code{\link[sq]{sq class}} documentation for details)}

\item{is_clean}{a \code{\link{logical}} value - if \code{TRUE}, letters will be drawn from
clean alphabet, if \code{FALSE} - ambiguous letters might be also generated}

\item{sd}{a positive \code{\link{numeric}} value; if specified, gives standard deviation of
length of generated sequences}

\item{use_gap}{- a \code{\link{logical}} value; if \code{TRUE}, sequences will be generated
with random gaps inside}
}
\value{
object of class \code{sq} with type as specified

Sequences are generated using \code{\link{sample}} function. There is no possibility of 
generating sequence of 0 length, even if \code{sd} is given. Letter '*' is not used 
in generating \strong{ami} sequences.
}
\description{
Generates \code{\link{sq}} object with specified number of sequences of given lenght 
and given type.
}
\examples{
# setting seed for reproducibility
set.seed(16)

# generating random sequences
random_sq(10, 10, "ami", TRUE)
random_sq(50, 8, "nuc", FALSE, sd = 3)
random_sq(6, 100, "ami", TRUE, use_gap = TRUE)
}
\seealso{
\code{\link{construct_sq}} \code{\link{sq}}
}
