% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{remove_na}
\alias{remove_na}
\title{Remove sequences containing NA values}
\usage{
remove_na(x, by_letter = FALSE)
}
\value{
A \code{\link{sq}} object with the same type as input type. Sequences not containing
any \code{\link[=sq]{NA}} values are left untouched.
}
\description{
Removes sequences containing ambiguous elements or removes \code{\link[=sq]{NA values}} 
from sequences in a \code{\link{sq}} object.
}
\details{
This function allows removal of sequences containing \code{\link[=sq]{NA}} values.
By default, whole sequences containing ambiguous elements are removed 
and \code{\link[=sq]{NULL}} (empty) sequences are introduced in their place. If 
\code{only_elements = TRUE} then only \code{\link[=sq]{NA}} values are removed 
from sequences in \code{sq} object. \code{\link[=sq]{NULL}} values (empty sequences) 
can be identified using \code{\link{is_null_sq}} function. 

\code{NA} may be introduced as a result of using functions like 
\code{\link{substitute_letters}} or \code{\link{bite}}. They also appear in sequences if
you are reading file using \code{\link{read_fasta}} or constructing \code{sq} object from
\code{\link{character}} vector with \code{\link{construct_sq}} in 
\code{\link[=fast-mode]{fast mode}} and there are letters in file or in strings other than
specified.
}
\seealso{
\code{\link{sq}} \code{\link{is_null_sq}} \code{\link{substitute_letters}}
\code{\link{bite}}
}
