% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\alias{reverse.sq}
\title{Reverse sequence}
\usage{
reverse(x, ...)

\method{reverse}{sq}(x, ..., NA_letter = getOption("tidysq_NA_letter"))
}
\arguments{
\item{x}{a \code{\link[=sq-class]{sq}} object.}

\item{...}{further arguments to be passed from or to other methods.}

\item{NA_letter}{[\code{character(1)}]\cr
A string that is used to interpret and display \code{NA} value in the
context of \code{\link[=sq-class]{sq class}}. Default value equals to
"\code{!}".}
}
\value{
A \code{\link[=sq-class]{sq}} object of the same type as input object but
each sequence is reversed.
}
\description{
Reverse given list of sequences.
}
\details{
The \code{reverse} function reverses each sequence in supplied 
\code{\link[=sq-class]{sq}} object (e.q. transforms "MIAANYTWIL" to "LIWTYNAAIM").
Empty sequences are left with no effect. This operation does not change 
the type of the input object nor its alphabet.

Since the function \code{reverse} returns a \code{\link[=sq-class]{sq}} object, the
\code{\link[=sq-print]{print}} function is implicitly called.
}
\seealso{
\code{\link[=sq-class]{sq}} \code{\link{clean}} \code{\link{sq-print}}
}
