% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm.R
\name{score_pssm}
\alias{score_pssm}
\title{Score a sequence against a PSSM using sum of positional scores.}
\usage{
score_pssm(sq, pssm)
}
\arguments{
\item{sq}{a \code{\link{sq}} object.}

\item{pssm}{a position specific scoring matrix}
}
\value{
A vector of scores, one for each input sequence.
}
\description{
Score a sequence against a PSSM using sum of positional scores.
}
\examples{
pssm1 <- empty_pssm(num_pos = 9, type = "nuc")
pssm1[1:9, 1:6] <- rnorm(54)
score_pssm(as.sq(c("GGGCTGGCG","CTTCAGACT")), pssm1)
}
