% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysq_options.R
\name{sq-options}
\alias{sq-options}
\title{Set options of package}
\usage{
\code{getOptions(option_name)}

\code{options(option_name = value)}
}
\arguments{
\item{option_name}{an option name to set or get}

\item{value}{a value to assing to an option}
}
\value{
in case of \code{getOptions} - value of the option
}
\description{
Set options of package
}
\details{
You can change default behaviour of package using one of following options:
\itemize{
\item tidysq_bite_na_action (default "warning") - specifies in which way to inform user
about biting sequences out of the range when using \code{\link{bite}},
\item tidysq_subsitute_letters_cln (default "warning") - specifies in which way to inform user
about droping \code{cln} subtype of \code{sq} while using \code{\link{substitute_letters}},
\item tidysq_typify_small_cap_let (default "warning") - specifies in which way to inform user
about merging lowercase and uppercase letters when typifying \code{sq} object,
\item tidysq_encode_no_given_action (default "warning") - specifies in which way to inform user
about encoding unspecified letters as \code{\link{NA}} if they do appear in sequences in
\code{\link{encode}},
\item tidysq_max_pillar_sq_width (default 15) - specifies pillar_shaft_sq width,
\item tidysq_max_print_sequences (default 10) - specifies maximum number of printed sequences
in \code{\link[sq:print.sq]{print sq}},
\item tidysq_colorful_sq_print (default \code{TRUE}) - specifies if to use colorful printing
(see \code{\link[sq:print.sq]{print sq}})
\item tidysq_na_print_char (default = "!") - specifies which character string to pring when 
\code{\link{NA}} values appear in sequences,
\item tidysq_fast_mode (default \code{FALSE}) - specifies if to work in fast mode (see 
\code{\link{fast-mode}})
}
}
\seealso{
\code{\link[base:option]{options}}
}
