% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysq_options.R
\name{tidysq-options}
\alias{tidysq-options}
\title{Set options of package}
\description{
You can get value of an option by calling \code{getOptions(option_name)} and set its value
by calling \code{options(option_name = value)}, where \code{option_name} is an option name 
(full list of this package included below) and \code{value} is a value to assing to an option.
}
\details{
You can change default behaviour of package using one of following options:
\itemize{
\item tidysq_bite_na_action (default "warning") - a \code{\link{character}} string specifying 
in which way to inform user about biting sequences out of the range when using 
\code{\link{bite}}; possible values: "error", "warning", "message", "none",
\item tidysq_subsitute_letters_cln (default "warning") - a \code{\link{character}} string 
specifying in which way to inform user
about droping \code{cln} subtype of \code{sq} while using \code{\link{substitute_letters}}; 
possible values: "error", "warning", "message", "none",
\item tidysq_typify_small_cap_let (default "warning") - a \code{\link{character}} string 
specifying in which way to inform user
about merging lowercase and uppercase letters when typifying \code{sq} object; possible 
values: "error", "warning", "message", "none",
\item tidysq_encode_no_given_action (default "warning") - a \code{\link{character}} string 
specifying in which way to inform user
about encoding unspecified letters as \code{\link{NA}} if they do appear in sequences in
\code{\link{encode}}; possible values: "error", "warning", "message", "none",
\item tidysq_max_pillar_sq_width (default 15) - an \code{\link{integer}} value specyfying 
pillar_shaft_sq width
\item tidysq_max_print_sequences (default 10) - an \code{\link{integer}} value specyfying 
maximum number of printed sequences
in \code{\link[=print.sq]{print sq}},
\item tidysq_colorful_sq_print (default \code{TRUE}) - a \code{\link{logical}} value if to 
use colorful printing
(see \code{\link[=print.sq]{print sq}})
\item tidysq_na_print_char (default = "!") -  a \code{\link{character}} string which character 
string to pring when 
\code{\link{NA}} values appear in sequences,
\item tidysq_fast_mode (default \code{FALSE}) - a \code{\link{logical}} value if to work in 
fast mode (see 
\code{\link{fast-mode}})
}

If value is not appropriate for given option, default value will be used.
}
\seealso{
\code{\link[base:option]{options}}
}
