% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unpack_nc_cnuc}
\alias{unpack_nc_cnuc}
\title{Unpack raw bytes of clean nucleotides sequence}
\usage{
unpack_nc_cnuc(PACKED, na_char)
}
\arguments{
\item{UNPACKED}{\code{raw} vector}

\item{ALPH_SIZE}{\code{integer}}
}
\description{
Unpack raw bytes of clean nucleotides sequence
}
