% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Save sq to fasta file}
\usage{
write_fasta(
  x,
  names,
  file,
  width = 80,
  NA_letter = getOption("tidysq_NA_letter")
)
}
\arguments{
\item{x}{[\code{sq}]\cr
An object this function is applied to.}

\item{file}{a \code{\link{character}} string indicating path to file to write into.}

\item{width}{a positive \code{\link{integer}} value informing about maximum number of 
characters to put in each line of file.}

\item{name}{a \code{\link{character}} vector of length equal to \code{sq} length.}
}
\description{
Writes \code{\link[=sq-class]{sq}} objects with their names to a fasta file.
}
