% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sq.R
\name{as.sq}
\alias{as.sq}
\title{Convert an object to sq}
\usage{
as.sq(x, ...)
}
\arguments{
\item{x}{a object of a class that supports conversion to \code{\link{sq}}.}

\item{...}{other arguments passed to the method.}
}
\value{
A \code{sq} object
}
\description{
This generic function takes an object of arbitrary type and returns a \code{\link{sq}} object 
as an output. Default implementation of the method throws an error - there needs to be 
implemented a method for specified class in order for function to work.
}
\details{
In \pkg{tidysq} only for \code{\link{character}} vector there is implemented a method 
which in fact works exactly like \code{\link{construct_sq}} - you can also pass other arguments
like those supported by \code{\link{construct_sq}}.
}
\examples{
# constructing an example in usual way
sq_1 <- construct_sq("CTGA")

# using a method for character
sq_2 <- as.sq("CTGA")

}
\seealso{
\code{\link{sq}} \code{\link{construct_sq}}
}
