% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bite.R
\name{bite}
\alias{bite}
\title{Subset sequences from sq}
\usage{
bite(sq, indices)
}
\arguments{
\item{sq}{\code{\link{sq}} object}

\item{indices}{\code{numeric} vector of subsequence indices to extract from
each sequence. Follows the normal R conventions for indexing vectors, 
including negative indices.}
}
\value{
\code{\link{sq}} object of the same type as input sq, where each 
element is a subsequence created by indexing corresponding sequence from 
input sq object with input indices.
}
\description{
Obtain subsequence from each sequence contained in the sq object with the 
same indices.
}
\details{
This function follows the normal R conventions, thus extracting 
indices not present in the sequence results in introducing NA values. 
Information about it is printed on console depending on value of option 
'tidysq_bite_na_action' - it can be either a warning (default), error, 
message or no information (you can check details in \link{sq-options}). 
NA values can be removed by using \link{remove_na} function.
}
\examples{
# creating object to work on:
sq <- construct_sq("ACTGCTG", "CTTAGA", "CCCT", "CTGAATGT")

# extracting first letter from each sequence:
bite(sq, 1)

# extracting first three letters from each sequence:
bite(sq, 1:3)

# extracting second, fourth, third and second letters:
bite(sq, c(2,4,3,2))

# extracting second to fifth letter - NA introduced:
bite(sq, 2:5)

# extracting all from first to twentieth - NA introduced:
bite(sq, 1:20)

# extracting all excluding first letter of sequence:
bite(sq, -1)

# extracting all excluding second and sixth letter of sequence:
bite(sq, c(-2, -6))


}
\seealso{
sq remove_na sq-options
}
