% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\title{Create complement sequence from nucseq object}
\usage{
complement(nucsq, is_dna = NULL)
}
\arguments{
\item{nucsq}{\code{\link{nucsq}} object of type 'nuc'.}

\item{is_dna}{logical indicating if entered sequence is DNA or RNA. If 
'is_dna' is \code{TRUE}, sequences cannot contain 'U' (uracil);
if is \code{FALSE}, sequences cannot contain 'T' (thymine). If 
\code{NULL} (default) the sequence type is not specified.}
}
\value{
\code{\link{nucsq}} object of the same type as input nucsq but 
built of complementary nucleotides to entered sequence.
}
\description{
Create complementary sequence from given nucleoide sequence 
according to complementarity.
}
\details{
This function allows obtaining  complement sequence which is created by 
matching elements (nucleotides) with complementarity to input nucsq object. If 
\code{is_dna = TRUE} entered sequence is DNA. If \code{is_dna = FALSE} entered 
sequence is RNA. By default the sequence type is not specified and function
needs to guess which type of sequence was entered. If sequence contain 'U' without 
'T' the type will be set to RNA. If sequence contain 'T' without 'U' the type will 
be set to dna. If both 'U' and 'T' are present in the sequence
or sequence contains 'A' (adenine), but doesn't contain neither 'T' nor 'U'
or logical indicating is typed wrong (i.e., when sequence contain 'U' and logical 
indicating is set to DNA) error will appear.
Both RNA and DNA sequences can be rewritten to complementary sequence.
}
\examples{
# Creating objects nuc_dna_sequence (with DNA sequences) and nuc_rna_sequence 
# (with RNA sequences) to work on:

nuc_dna_sequence <- construct_sq(c("ACTGCTG", "CTTAGA", 
                                   "CCCT", "CTGAATGT"), type = "nuc")
nuc_rna_sequence <- construct_sq(c("ACUGCUG", "CUUAGA", 
                                   "CCCU", "CUGAAUGU"), type = "nuc")
                                   
                                   
# Sequences are now ready to be used as matrices for creating complementary sequences.
# Get an overview of the sequences:
summary(nuc_dna_sequence)
summary(nuc_rna_sequence)

# In the following the complement function is used to make a PCR (Polymerase Chain Reaction)
# primers. Every sequence will be rewritten to its complementary equivalent as 
# following example: AAATTTGGG to TTTAAACCC.
 
# creating complementary sequences with defined sequence type:
complement(nuc_dna_sequence, is_dna = TRUE)
complement(nuc_rna_sequence, is_dna = FALSE)

# creating complementary sequences without defined sequence type:
complement(nuc_dna_sequence)
complement(nuc_rna_sequence)

# Each sequence from  nuc_dna_sequence and nuc_rna_sequence object have now own 
# complementary equivalent, which can be helpful during constructing PCR primers.
 
  

}
\seealso{
sq clnsq nucsq
}
