% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\alias{complement_dna}
\alias{complement_rna}
\title{Create complement sequence from dnasq or rnasq object}
\usage{
complement(x, NA_letter = getOption("tidysq_NA_letter"), ...)

complement_dna(x)

complement_rna(x)
}
\arguments{
\item{x}{a \code{\link{sq}} object of type \strong{dna} or \strong{rna}.}
}
\value{
\code{sq} object of the same type as input \code{dnasq} (\strong{dna})
or \code{rnasq} (\strong{rna}) but built of complementary nucleotides to entered
sequence.
}
\description{
Creates the complementary sequence from a given RNA or DNA 
 nucleotide sequence. The function differentiates between RNA and DNA sequences.
}
\details{
This function allows to get complement sequence, which is created by 
matching elements (nucleotides) with complementary to input dnasq or rnasq object.
Whether 'U' (uracil) or 'T' (thymine) is used depends on the class of the sq object.

Functions \code{complement_dna} and \code{complement_rna} are provided as a safe
way of limiting classes \code{complement} function is used on.
}
\seealso{
\code{\link{sq}}
}
