% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\alias{complement_dna}
\alias{complement_rna}
\title{Create complement sequence from dnasq or rnasq object}
\usage{
complement(sq)

complement_dna(sq)

complement_rna(sq)
}
\arguments{
\item{nucsq}{a \code{\link{sq}} object of type \strong{dna} or \strong{rna}.}
}
\value{
\code{sq} object of the same type as input \code{dnasq} (\strong{dna})
or \code{rnasq} (\strong{rna}) but built of complementary nucleotides to entered
sequence.
}
\description{
Creates the complementary sequence from a given RNA or DNA 
 nucleotide sequence. The function differentiates between RNA and DNA sequences.
}
\details{
This function allows to get complement sequence, which is created by 
matching elements (nucleotides) with complementary to input dnasq or rnasq object.
Whether 'U' (uracil) or 'T' (thymine) is used depends on the class of the sq object.

Functions \code{complement_dna} and \code{complement_rna} are provided as a safe
way of limiting classes \code{complement} function is used on.
}
\examples{
# Creating objects dna_sequence (with DNA sequences) and rna_sequence 
# (with RNA sequences) to work on:

dna_sequence <- construct_sq(c("ACTGCTG", "CTTAGA", "CCCT", "CTGAATGT"),
                             type = "dna")
rna_sequence <- construct_sq(c("ACUGCUG", "CUUAGA", "CCCU", "CUGAAUGU"),
                             type = "rna")


# In the following the complement function is used to make a PCR (Polymerase Chain Reaction)
# primers. Every sequence will be rewritten to its complementary equivalent as 
# following example: AAATTTGGG to TTTAAACCC.
 
# creating complementary sequences with the basic function and using wrappers:
complement(dna_sequence)
complement_dna(dna_sequence)

complement(rna_sequence)
complement_rna(rna_sequence)

# Each sequence from dna_sequence and rna_sequence object have now an own
# complementary equivalent, which can be helpful during constructing PCR primers.

}
\seealso{
\code{\link{sq}}
}
