% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sq.R
\name{==.sq}
\alias{==.sq}
\title{Compare sq object}
\usage{
\method{==}{sq}(e1, e2)
}
\arguments{
\item{x1}{\code{\link{sq}} object.}

\item{x2}{an object to compare with \code{\link{sq}}.}
}
\description{
Compares input \code{\link{sq}} object with another given.
}
\details{
\code{`==`} checks if the input object is sequence, if yes, converts
it to chracters and checks whether given object can be compared with
\code{\link{sq}} object. If given sequence consists lowercase, the function
rewrites them into capital ones.
}
\examples{

# Creating sq object to work on:
sq <- construct_sq(c("ACTGCTG", "CTTAGA", 
                     "CCCT", "CTGAATGT"), type = "nuc")
nuc_dna_sequence <- construct_sq(c("ACTGCTG", "CTTAGA", 
                                   "GGAA", "ATGAACGT"), type = "nuc")                              
# Get an overview of the sequences:
summary(sq)
summary(nuc_dna_sequence)

# Comparing object nuc_dna_sequence to \\code{\\link{sq}}:

`==`(\\code{\\link{sq}},nuc_dna_sequence)

# Also comparing object nuc_dna_sequence to \\code{\\link{sq}}:

 \\code{\\link{sq}} == nuc_dna_sequence
                                                                    
}
