% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sq.R
\name{export_sq}
\alias{export_sq}
\title{Export sq objects into other formats}
\usage{
export_sq(sq, export_format, name = NULL, ...)
}
\arguments{
\item{sq}{a \code{\link{sq}} object.}

\item{export_format}{a \code{\link{character}} string indicating package and the destination 
class; it should be one of the following: "seqinr::SeqFastaAA", "ape::AAbin", 
"Biostrings::AAStringSet", "seqinr::SeqFastadna", "ape::DNAbin", "Biostrings::DNAStringSet".}

\item{name}{a \code{\link{character}} vector of length equal to \code{sq} length.}

\item{...}{- additional arguments passed to the function.}
}
\description{
Convert object of class \code{\link{sq}} to another class from another package. Currently 
supported packages are \pkg{ape} with its formats (\code{AAbin} and \code{DNAbin}),
\pkg{Bioconductor} (\code{AAStringSet}, \code{DNAStringSet}) and
\pkg{seqinr} (\code{SeqFastaAA}, \code{SeqFastadna}).
}
\examples{
sq_ami <- construct_sq(c("MVVGL", "LAVPP"))
export_sq(sq_ami, "ape::AAbin")
export_sq(sq_ami, "Biostrings::AAStringSet", c("one", "two"))
export_sq(sq_ami, "seqinr::SeqFastaAA")

sq_dna <- construct_sq(c("TGATGAAGCGCA", "TTGATGGGAA"))
export_sq(sq_dna, "ape::DNAbin", name = c("one", "two"))
export_sq(sq_dna, "Biostrings::DNAStringSet")
export_sq(sq_dna, "seqinr::SeqFastadna")
}
\seealso{
\code{\link{sq}} \code{\link{import_sq}}
}
