% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sq.R
\name{export_sq}
\alias{export_sq}
\title{Export sq objects into other formats}
\usage{
export_sq(x, export_format, name = NULL, ...)
}
\arguments{
\item{x}{[\code{sq}]\cr
An object this function is applied to.}

\item{export_format}{[\code{character(1)}]\cr
A string indicating desired class (with specified package for unambiguity).}

\item{name}{[\code{character}]\cr
Vector of sequence names. Must be of the same length as \code{sq} object.
 Can be \code{NULL}.}

\item{...}{further arguments to be passed from or to other methods.}
}
\description{
Converts object of class \code{\link[=sq-class]{sq}} to a class
from another package. Currently supported packages are \pkg{ape},
\pkg{Bioconductor} and \pkg{seqinr}. For exact list of supported classes and
resulting types, see details.
}
\details{
Currently supported formats are as follows (grouped by \code{sq} types):
\itemize{
\item \strong{ami}:
 \itemize{
 \item \code{"ape::AAbin"}
 \item \code{"Biostrings::AAString"}
 \item \code{"Biostrings::AAStringSet"}
 \item \code{"seqinr::SeqFastaAA"}
 }
\item \strong{dna}:
 \itemize{
 \item \code{"ape::DNAbin"}
 \item \code{"Biostrings::DNAString"}
 \item \code{"Biostrings::DNAStringSet"}
 \item \code{"seqinr::SeqFastadna"}
 }
\item \strong{rna}:
 \itemize{
 \item \code{"Biostrings::RNAString"}
 \item \code{"Biostrings::RNAStringSet"}
 }
}
}
\examples{
# DNA and amino acid sequences can be exported to most packages
sq_ami <- sq(c("MVVGL", "LAVPP"), alphabet = "ami_bsc")
export_sq(sq_ami, "ape::AAbin")
export_sq(sq_ami, "Biostrings::AAStringSet", c("one", "two"))
export_sq(sq_ami, "seqinr::SeqFastaAA")

sq_dna <- sq(c("TGATGAAGCGCA", "TTGATGGGAA"), alphabet = "dna_bsc")
export_sq(sq_dna, "ape::DNAbin", name = c("one", "two"))
export_sq(sq_dna, "Biostrings::DNAStringSet")
export_sq(sq_dna, "seqinr::SeqFastadna")

# RNA sequences are limited to Biostrings
sq_rna <- sq(c("NUARYGCB", "", "DRKCNYBAU"), alphabet = "rna_ext")
export_sq(sq_rna, "Biostrings::RNAStringSet")

# Biostrings accept single sequences as well
export_sq(sq_dna[1], "Biostrings::DNAString")

}
\seealso{
\code{\link[=sq-class]{sq class}}

Functions from I/O module:
\code{\link{import_sq}()},
\code{\link{random_sq}()},
\code{\link{sq}()}
}
\concept{io_functions}
