% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motifs}
\alias{find_motifs}
\title{Find given motifs}
\usage{
find_motifs(x, names, motifs, NA_letter = getOption("tidysq_NA_letter"), ...)
}
\arguments{
\item{x}{a \code{\link{sq}} object.}

\item{motifs}{a \code{character} vector of motifs to be searched for.}

\item{name}{a non-\code{NULL} \code{character} vector without \code{\link{NA}} values, 
containing names of the sequences in the sq. It has to be of the same length 
as the \code{sq}.}
}
\value{
A \code{\link[tibble]{tibble}} with following columns:
 \item{name}{name of the sequence}
 \item{sq}{sequence}
 \item{sought}{sought motif}
 \item{found}{motif found in a sequence, may differ from sought if a motif
 contained ambiguous letters}
 \item{start}{position of motif start}
 \item{end}{position of motif end}
}
\description{
Find all given motifs in all sequences and return their 
positions.
}
\details{
This function allows search of a given motif or motifs in the \code{sq} 
object. It returns all motifs found with their start and end positions 
within a sequence.
}
\section{Allowed and forbidden letters and characters details}{

Note if a sq object contains characters: ^$?=()\.|+*{}[] in its alphabet, 
search for motifs cannot be performed and an error will be displayed (with 
exception of sq objects of type ami - there is '*' letter in their alphabet
and it can be contained in sought motif). To search for motifs with those 
characters, you have to replace them first using 
\code{\link{substitute_letters}}. 

If sq objects of type \strong{ami}, \strong{dna} and \strong{rna}, motifs have to
consist of upper case letters from amino acid, DNA and RNA alphabets respectively.
Use of lower case letters will return an error. Two additional characters 
are allowed: '^' and '$' indicating the beginning and the end of a sequence 
respectively. Moreover, notice that '*' character may be used in amino acid 
motifs, as it is a part of the amino acid alphabet. If a motif contains 
ambiguous letters, all possible matches will be searched for. For example the 
amino acid motif "MAJ" (where "J" is an ambiguous letter indicating L or I) will 
find the motifs: "MAJ", "MAL" and "MAI". 

Detailed list of all letters corresponding to each ambiguous letter may be found at
\code{\link{aminoacids_df}} and \code{\link{nucleotides_df}}.
}

\examples{
# Creating objects to work on:
sq_ami <- construct_sq(c("AGNTYIKFGGAYTI", "MATEGILIAADGYTWIL", 
                         "MIPADHICAANGIENAGIK"), type = 'ami')
sqtbl <- read_fasta(system.file(package = "tidysq", "example_aa.fasta"), 
                    type = "ami")

# Find motif of two alanines followed by aspartic acid or asparagine 
# ('AAB' motif will match 'AAB', 'AAD' and 'AAN'):
find_motifs(sq_ami, c("sq1", "sq2", "sq3"), "AAB")

# Find motif 'VHH' at the beginning of sequences:
find_motifs(sqtbl[["sq"]], sqtbl[["name"]], "^VHH")

# Find motif 'DPGS' at the end of sequences:
find_motifs(sqtbl[["sq"]], sqtbl[["name"]], "DPGS$")

# Find multiple motifs:
find_motifs(sqtbl[["sq"]], sqtbl[["name"]], c("^LIV", "XXKK", "EN$"))

}
\seealso{
\code{\link{sq}} \code{\link{substitute_letters}} \code{\link{\%has\%}}
}
