% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sq_type.R
\name{get_sq_alphabet}
\alias{get_sq_alphabet}
\title{Get alphabet of given sq object.}
\usage{
get_sq_alphabet(sq)
}
\arguments{
\item{sq}{a \code{\link{sq}} object to be recognized.}
}
\value{
A character vector of letters of the alphabet.
}
\description{
Function returns amino acid, DNA, RNA or atypical alphabet based on a \code{\link{sq}} 
object type.
}
\details{
This function allows returning alphabet of \code{sq} object which is a character
or numeric vector. The function reads provided \code{\link{sq}} object and determines,
which kind of sequences user assigned to a \code{\link{sq}} object (DNA, RNA, amino acid,
atypical or encoded one).

If \code{\link{sq}} has \strong{ami} type and \strong{cln} subtype, the function returns
the set of 20 aminoacids, the gap (-) and the stop codon (*) letter. If \strong{cln} is
missing, ambiguous letters are included as well.

If a \code{\link{sq}} contains \strong{dna} or \strong{rna} sequences and \strong{cln}
subtype, the function returns a set of respective 4 nucleotides with gap (-) element.
If \strong{cln} is missing, ambiguous letters are included as well.

If \code{\link{sq}} type is \strong{unt} or \strong{atp}, the function returns a list of 
letters present in sequences of a \code{\link{sq}} object.

If type is \strong{enc}, a numeric vector of values encoded for letters is returned
(see \code{\link{encode}}).

The details about amino acid and nucleotide alphabets can be checked in
\code{\link{aminoacids_df}} and \code{\link{nucleotides_df}} respectively. General
information about alphabets and types of \code{sq} objects can be found in \code{\link{sq}}
class documentation.
}
\examples{
# Creating an object to work on:
sq_dna <- construct_sq(c("ACGATTAGACG","GGATA"), type = "dna")
sq_amino_acids <- construct_sq(c("MMVTAAV"), type = "ami")
sq_untyped <- construct_sq(c("ACGA&&TTAGACG&"), type = "unt")

# Testing DNA sq object with defined type:
get_sq_alphabet(sq_dna)

# Testing amino acid sq object with defined type:
get_sq_alphabet(sq_amino_acids)

# Testing nucleotide sq object without defined type:
get_sq_alphabet(sq_untyped)

  
}
\seealso{
\code{\link{sq}} \code{\link{construct_sq}} \code{\link{encode}}
}
