% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-methods.R
\name{get_sq_lengths}
\alias{get_sq_lengths}
\title{Get lengths of sequences in sq object}
\usage{
get_sq_lengths(x)
}
\arguments{
\item{x}{an \code{\link[=sq-class]{sq}} object.}
}
\value{
A \code{\link{numeric}} vector, where each element gives length of according 
sequence from \code{\link[=sq-class]{sq}} object.
}
\description{
Function counts number of elements in each sequence in given \code{\link[=sq-class]{sq}} object.
}
\details{
This function allows returning numeric vector of lengths of sequences from
\code{\link[=sq-class]{sq}} object. The numeric vector is as long as number of sequences present
in \code{\link[=sq-class]{sq}} object.
The function counts elements in all types of sequences.
}
\seealso{
\code{\link[=sq-class]{sq}} \code{\link{construct_sq}}
}
