% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_funs.R
\name{get_sq_type}
\alias{get_sq_type}
\title{Checks type of the sequence}
\usage{
get_sq_type(sq)
}
\arguments{
\item{sq}{\code{\link{sq}} object to be checked.}
}
\value{
a type string and its name defined in \code{\link{sq}} object.
}
\description{
Function is checking which type of sequence contains \code{\link{sq}} object.
}
\details{
This function returns type of sequence from \code{\link{sq}} object.
If the type of sequence is "nuc", "ami" or "unt"  function returns 
(nucsq) "nuc", (amisq) "ami" or (untsq) "unt" respectivetly.
}
\examples{
# Creating an object to work on:
sq_nucleotides <- construct_sq(c("ACGATTAGACG","GGATA"), type = "nuc")
sq_amino_acids <- construct_sq(c("MMVTAAV"), type = "ami")
sq_unt <- construct_sq(c("MMVTAAVXX"), type = "unt")

# Gettig sequence type from nucleotide sq object:
get_sq_type(sq_nucleotides)

# Gettig sequence type from amino acid  sq object:
get_sq_type(sq_amino_acids)

# Gettig sequence type from unt sq object:
get_sq_type(sq_unt)


}
\seealso{
sq construct_sq
}
