% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{\%has\%}
\alias{\%has\%}
\title{Test sq object for presence of given motifs}
\usage{
x \%has\% y
}
\arguments{
\item{x}{a \code{\link{sq}} object to be tested.}

\item{y}{a \code{character} vector of motifs to be searched for.}
}
\value{
A \code{\link{logical}} vector of the same length as input \code{sq}, indicating 
which elements contain all given motifs
}
\description{
Test if elements of a \code{\link{sq}} object contain given motifs
}
\details{
This function allows testing if elements of a \code{sq} object contain 
a given motif or motifs, which includes start sequences and stop codons. It returns 
a \code{logical} for every element of the \code{sq} object - \code{TRUE} if it 
contains the motif and \code{FALSE} otherwise. When multiple motifs are searched, 
\code{TRUE} will be returned only for sequences that contain all the given motifs. 

This function only indicates if a motif is present within a sequence, to 
find all motifs, and their positions within sequences use 
\code{\link{find_motifs}}.
}
\section{Allowed and forbidden letters and characters details}{

Note: if a sq object contains characters: ^$?=()\.|+*{}[] in its alphabet, 
search for motifs cannot be performed and an error will be displayed (with 
exception of sq objects of type ami - in their alphabet there is '*' letter 
and it can be contained in sought motif). To search for motifs with those 
characters, you have to replace them first using 
\code{\link{substitute_letters}}. 

In case of sq objects of type \strong{ami}, \strong{dna} and \strong{rna},
motifs have to consist of upper case letters from amino acid, DNA and RNA alphabets
respectively. Use of lower case letters will return an error. Two additional
characters are allowed: '^' and '$' indicating the beginning and the end of a sequence 
respectively. Moreover, notice that '*' character may be used in amino acid 
motifs, as it is a part of the amino acid alphabet. If a motif contains 
ambiguous letters, all possible matches will be searched for, e.g., amino 
acid motif "MAJ" (where "J" is an ambiguous letter indicating L, or I) will 
find motifs: "MAJ", "MAL" and "MAI". 

Detailed list of all letters corresponding to each ambiguous letter may be found at
\code{\link{aminoacids_df}} and \code{\link{nucleotides_df}}.
}

\seealso{
\code{\link{sq}} \code{\link{substitute_letters}} \code{\link{find_motifs}}
}
