% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sq.R
\name{import_sq}
\alias{import_sq}
\title{Import sq objects from other objects}
\usage{
import_sq(object, ...)
}
\arguments{
\item{object}{- an object of one of classes: \code{AAbin}, \code{DNAbin}, \code{AAStringSet}, 
\code{DNAStringSet}, \code{SeqFastaAA}, \code{SeqFastadna}.}
}
\value{
A \code{\link[tibble]{tibble}} with \code{sq} column of \code{\link{sq}} type 
representing the same 
sequences as given object; the object has a type corresponding to the input type; if given
sequences had names, output \code{\link[tibble]{tibble}} has also another column 
\code{name} with those names
}
\description{
Creates \code{\link[=sq]{sq object}} from \code{object} of class from another package.
Currently supported packages are \pkg{ape} with its formats (\code{AAbin} and \code{DNAbin}),
\pkg{Bioconductor} (\code{AAStringSet}, \code{DNAStringSet}) and
\pkg{seqinr} (\code{SeqFastaAA}, \code{SeqFastadna}).
}
\details{
Providing object of class other than specified will result in error.
}
\examples{
## ape example
library(ape)
ape_dna <- as.DNAbin(list(one = c("C", "T", "C", "A"), two = c("T", "G", "A", "G", "G")))
import_sq(ape_dna)

## Biostrings example
library(Biostrings)
Biostrings_dna <- DNAStringSet(c(one = "CTCA", two = "TGAGG"))
import_sq(Biostrings_dna)

## seqinr example
library(seqinr)
seqinr_dna <- as.SeqFastadna(list(one = c("C", "T", "C", "A"), 
                                  two = c("T", "G", "A", "G", "G")))
import_sq(seqinr_dna)

}
\seealso{
\code{\link{export_sq}} \code{\link{sq}}
}
