% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-methods.R
\name{is.sq}
\alias{is.sq}
\alias{is.sq_ami_bsc}
\title{Check if object has specified type}
\usage{
is.sq(x)

is.sq_ami_bsc(x)
}
\arguments{
\item{x}{an object to be checked.}
}
\value{
A \code{\link{logical}} value - \code{TRUE} if \code{x} has specified type, \code{FALSE}
otherwise.
}
\description{
Function to checks if object is a \code{\link{sq}} object without specifying type or
if it is a \code{\link{sq}} object with specific type.
}
\details{
These function does not only check objects classes - they also check if their format is 
correct (e.g. if they have alphabet parameter, if they have exactly one type, if they are
list of raws, etc. - to see details, how does \code{sq} object look like under the hood, 
read \code{\link[=sq]{sq class}} manual).
}
\seealso{
\code{\link{sq}}
}
