% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_null_sq.R
\name{is_null_sq}
\alias{is_null_sq}
\title{Test if a sequence is empty}
\usage{
is_null_sq(x)
}
\arguments{
\item{sq}{a\code{\link{sq}} object to be tested}
}
\value{
A logical vector of the same length as input sq, 
indicating which elements are \code{NULL sq}, i.e., an empty sequence
of length 0.
}
\description{
Test a sq object for presence of empty sequences
}
\details{
This function allows identification of empty sequences (that
have length 0) represented by the \code{NULL sq} values in the sq object. 
It returns a logical for every element of the sq object - \code{TRUE} if
its value is \code{NULL sq} and \code{FALSE} otherwise. \code{NULL sq} 
values may be introduced as a result of \code{\link{clean}} and 
\code{\link{remove_na}} functions. The first one replaces sequences 
containing ambiguous elements with \code{NULL sq} values, whereas the 
latter replaces \code{NA} values (which may be introduced by 
\code{\link{bite}}) with \code{NULL sq}.
}
\seealso{
\code{\link{sq}} \code{\link{clean}}
}
