% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sq.R
\name{random_sq}
\alias{random_sq}
\title{Generate random sequences}
\usage{
random_sq(n, len, alphabet, sd = NULL, use_gap = FALSE)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
A number of sequences to generate - must be non-negative.}

\item{len}{[\code{integer(1)}]\cr
Length of each sequence if \code{sd} not specified and mean length of
sequences if \code{sd} specified - must be non-negative.}

\item{alphabet}{[\code{character}]\cr
If provided value is a single string, it will be interpreted as type (see
details). If provided value has length greater than one, it will be treated
as atypical alphabet for \code{sq} obejct and \code{sq} type will be
\code{atp}.}

\item{sd}{[\code{integer(1)}]\cr
If specified, gives standard deviation of length of generated sequences -
must be non-negative.}

\item{use_gap}{[\code{logical(1)}]\cr
If \code{TRUE}, sequences will be generated with random gaps inside
(commonly denoted as "\code{-}").}
}
\value{
An object of class \code{sq} with type as specified.
}
\description{
Generates an \code{\link[=sq-class]{sq}} object with specified
number of sequences of given length and alphabet.
}
\details{
Letter '*' is not used in generating \strong{ami} sequences. If parameter
\code{sd} is passed, then all generated negative values are replaced with 0s.
}
\seealso{
Functions from I/O module
\code{\link{sq}()}
}
\concept{io_functions}
