% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ambiguous.R
\name{remove_ambiguous}
\alias{remove_ambiguous}
\title{Remove sequences that contain ambiguous elements}
\usage{
remove_ambiguous(x, by_letter = FALSE, ...)
}
\arguments{
\item{x}{}

\item{...}{further arguments to be passed from or to other methods.}

\item{NA_letter}{[\code{character(1)}]\cr
A string that is used to interpret and display \code{NA} value in the
context of \code{\link[=sq-class]{sq class}}. Default value equals to
"\code{!}".}
}
\description{
Remove sequences that contain ambiguous elements
}
