% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Reverse sequence}
\usage{
reverse(sq)
}
\arguments{
\item{sq}{\code{\link{sq}} object}
}
\value{
\code{\link{sq}} object of the same type as input sq, where each element is reversed.
}
\description{
Reverse given list of sequences.
}
\examples{
# Reverse just one sequence
reverse(construct_sq("ACTAGAGTGATAGA", type = "nuc"))

# Reverse list of sequences
reverse(construct_sq(c("fafasfasfFSA", "ygagayagfa", "adsDaf"), type = "ami"))

# Reverse list of uncleaned sequences
reverse(construct_sq(c("PASJIFEHF", "hvfisxxx", "xxxer")))

# Reverse list of cleaned sequences - only_elements = FALSE
reverse(clean(construct_sq(c("fafasfasfFSA", "ygagayagfa", "adxxaf", "xxx"), type = "ami")))

# Reverse list of cleaned sequences - only_elements = TRUE
reverse(clean(construct_sq(c("fafasfasfFSA", "ygagayagfa", "adxxaf", "xxx"), type = "ami"), only_elements = TRUE))

}
