% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Reverse sequence}
\usage{
reverse(sq)
}
\arguments{
\item{sq}{sq object}
}
\value{
reversed sq object (with preserved type)
}
\description{
Reverse given list of sequences.
}
\examples{
reverse(construct_sq("ACTAGAGTGATAGA", type = "nuc"))
reverse(construct_sq(c("fafasfasfFSA", "ygagayagfa", "adsDaf"), type = "ami"))

}
