% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_sq.R
\name{sq}
\alias{sq}
\title{sq: class for keeping biological sequences tidy}
\description{
An object of class \strong{sq} represents a list of biological sequences. It is the 
main internal format of the \pkg{tidysq} package and most functions operate on it. 
The storage method is memory-optimized so that objects require as little memory
as possible (details below).
}
\section{Construction/reading/import of sq objects}{

There are multiple ways of obtaining \code{sq} objects:
\itemize{
\item constructing from a \code{\link{character}} vector with \code{\link{construct_sq}},
\item constructing from another object with \code{\link{as.sq}} method,
\item reading from the fasta file with \code{\link{read_fasta}},
\item importing from a format of other package like \pkg{ape} or
\pkg{Biostrings} with \code{\link{import_sq}}.
}

\strong{Important note:} A manual assignment of a class \code{sq} to an object is
\strong{strongly discouraged} - due to the usage of low-level functions for
bit packing such assignment may lead to calling one of those functions during
operating on object or even printing it which can cause a crash of R session and,
in consequence, loss of data.
}

\section{Export/writing of sq objects}{

There are multiple ways of saving \code{sq} objects or converting them into
other formats:
\itemize{
\item converting into a character vector with \code{\link[=as.character.sq]{as.character}} 
method,
\item converting into a character matrix (or numeric matrix, in case of \strong{enc}) 
with \code{\link[=as.matrix.sq]{as.matrix}} method,
\item converting into a list of numerics with \code{\link{encsq_to_list}} (only
for \strong{enc} \code{sq}),
\item saving as fasta file with \code{\link{write_fasta}},
\item exporting into a format of other package like \code{ape} or
\code{Biostrings} with \code{\link{export_sq}}.
}
}

\section{Types of sq}{

This package is meant to handle both amino acid, DNA and RNA sequences, 
thus there is need to differentiate \code{sq} objects that keep them. In 
addition, there are special types for handling non-standard sequence 
formats and encodings.

Each \strong{sq} object has exactly one of \strong{types}:
\itemize{
\item \strong{ami} - (\emph{amino acids}) represents a list of sequences of amino acids
(peptides or proteins),
\item \strong{dna} - (\emph{DNA}) represents a list of DNA sequences,
\item \strong{rna} - (\emph{RNA}) represents a list of RNA sequences (together with
\strong{dna} above often collectively called "nucleotide sequences"),
\item \strong{unt} - (\emph{untyped}) represents a list of sequences that do not have 
specified type. They are mainly result of reading sequences from a file that 
contains some letters that are not in standard nucleotide or amino acid alphabets 
and user has not specified them explicitly. They should be converted to \strong{ami} ,
\strong{dna} or \strong{rna} sequences (using functions like \code{\link{substitute_letters}}
or \code{\link{typify}}).
\item \strong{atp} - (\emph{atypical}) represents sequences that have an alphabet
different from standard \strong{ami}, \strong{dna} or \strong{rna} alphabets - similarly
to \strong{unt}, but user has explicitly informed about it. They are
result of constructing sequences or reading from file with specifying 
\code{non_standard} parameter (for details see \code{\link{read_fasta}}
and \code{\link{construct_sq}}). They are also result of using function
\code{\link{substitute_letters}} - users can use this for example to simplify 
an alphabet and to replace a few letters by one.
\item \strong{enc} - (\emph{encoded}) represents list of sequences that have been
encoded with function \code{\link{encode}} where each letter is assigned with 
numeric value.
}

Additionally, there is a special subtype \strong{cln} (standing for \emph{clean}).
Only \strong{ami}, \strong{dna} or \strong{rna} \code{sq} objects may have this subtype.
It indicates that sequences do not contain ambiguous letters (see "alphabets" section below).

\code{sq} object type is printed when using overloaded method 
\code{\link[=print.sq]{print}}. It can be also checked by using \code{\link{get_sq_type}}
}

\section{Alphabet}{

Each \code{sq} object have an \strong{alphabet} associated with it. Alphabet is
a set of possible \strong{letters} that can appear in sequences contained in object.
Alphabet is kept mostly as a character vector, where each element represents one
\strong{letter}.

\code{sq} objects of type \strong{ami}, \strong{dna} or \strong{rna} have fixed alphabets
(depending also if object has \strong{cln} subtype). In other words, if two \code{sq}
objects have exactly the same type - \strong{ami}, \strong{dna} or \strong{rna} - and
either both have or both do not have a \strong{cln} subtype, they are ensured to have
the same alphabets.

Here are listed alphabets for these types:
\itemize{
\item \strong{ami} \strong{cln} - ACDEFGHIKLMNPQRSTVWY-*
\item \strong{ami}, (not \strong{cln}) - ABCDEFGHIJKLMNOPQRSTUVWXYZ-*
\item \strong{dna} \strong{cln} - ACGT-
\item \strong{dna}, (not \strong{cln}) - ACGTWSMKRYBDHVN-
\item \strong{rna} \strong{cln} - ACGU-
\item \strong{rna}, (not \strong{cln}) - ACGUWSMKRYBDHVN-
}

To see details of these alphabets see \code{\link{aminoacids_df}} and 
\code{\link{nucleotides_df}}.

Other types of \code{sq} objects are allowed to have different alphabets. Having an alphabet
exactly identical to one of those above does not automatically indicate that the type of
the sequence is one of those - e. g., there might be \strong{unt} \code{sq} that has an alphabet
identical to \strong{ami} \strong{cln} alphabet. To set the type, one should use the
\code{\link{typify}} function.

The purpose of co-existence of \strong{unt} and \strong{atp} alphabets is the 
fact that although there is a standard for format of \emph{fasta} files, sometimes 
there are other types of symbols, which do not match the standard. Thanks to these types, 
tidysq can import files with customized alphabets can be imported. Moreover, a user 
may want to group amino acids with similar properties (e. g., for machine learning) 
and replace the longer alphabet with symbols of groups. To check details, see 
\code{\link{read_fasta}}, \code{\link{construct_sq}} and \code{\link{substitute_letters}}.

All of the types: \strong{ami}, \strong{dna}, \strong{rna}, \strong{atp}, \strong{unt}
have alphabets that are character vectors while \strong{enc} objects have alphabets that
are numeric vectors. They are result of \code{\link{encode}} function, see its manual
for details.

\strong{Important note:} in \strong{atp} alphabets there is possibility of appearance of
letters that consists of more than one character - this functionality is provided in
order to handle situations like post-translational modifications, (e.g., using "mA" to 
indicating methylated alanine).

\strong{Important note:} alphabets of \strong{atp} and \strong{unt} \code{sq} objects
are case sensitive. Thus, in their alphabets both lowercase and uppercase letters can
appear simultaneously and they are treated as different characters. Alphabets of
\strong{dna}, \strong{rna} and \strong{ami} objects are always uppercase and all functions converts
other parameters to uppercase when working with \strong{dna}, \strong{rna} or \strong{ami} - e.g.,
\code{\link{\%has\%}} operator converts lower letters to upper when searching for motifs in
\strong{dna}, \strong{rna} or \strong{ami} object.

\strong{Important note:} maximum length of an alphabet is \strong{30 letters}. You are
not allowed to read fasta files or construct from character vectors that have more
than 30 distinct characters in sequences (with exception of reading or constructing
\strong{ami}, \strong{dna} or \strong{rna} objects - during their construction lowercase
letters are automatically converted to uppercase).

You can obtain an alphabet of the \code{sq} object using the \code{\link{get_sq_alphabet}}
function. You can check which letters are invalid (are not represented in standard
amino acid or nucleotide alphabet) in each sequence of given \code{sq} object of type
\strong{unt} or \strong{atp} by using \code{\link{get_invalid_letters}}. You can
substitute one letter with another using \code{\link{substitute_letters}}.
}

\section{Missing/Not Available values}{

There is a possibility of introducing \code{\link{NA}} values into sequences. \code{NA} value
does not represents gap (which are represented by \code{'-'}) or wildcard elements 
(\code{N} in the case of nucleotides and \code{'X'} in the case of amino acids), but is used 
as a representation of an empty position or invalid letters (not represented in nucleotide or 
amino acid alphabet).
\code{NA} does not belong to any alphabet (with exception of \strong{enc} objects where some of
letters might be \code{NA}, see \code{\link{encode}} for details). It is printed as
'!' and, thus, it is highly unrecommended to use '!' as special letter in \strong{atp}
sequences (but print character can be changed in options, see \code{\link{tidysq-options}}).

\code{NA} might be introduced by:
\itemize{
\item reading fasta file with non-standard letters in
\code{\link[=fast-mode]{fast mode}} with \code{\link{read_fasta}},
\item replacing a letter with \code{NA} value with \code{\link{substitute_letters}},
\item subsetting sequences beyond their lengths with \code{\link{bite}}.
}

An user can convert sequences that contain \code{NA} values into \code{NULL}
sequences with \code{\link{remove_na}}.
}

\section{NULL (empty) sequences}{

\code{NULL} sequence is a sequence of length 0.

\code{NULL} sequences might be introduced by:
\itemize{
\item constructing \code{sq} object from character string of length zero with
\code{\link{construct_sq}},
\item using the \code{\link{clean}} function,
\item using the \code{\link{remove_na}} function,
\item subsetting \code{sq} object with \link[=sq-extract]{extract operator}
}
}

\section{Storage format}{

\code{sq} object is, in fact, \strong{list of raw vectors}. The fact that it is list
implies that an user can concatenate \code{sq} objects using \code{\link[=c.sq]{c}} method
and subset them using \code{\link[=sq-extract]{extract operator}}. Alphabet is kept
as an attribute of the object. 

Raw vectors are the most efficient way of storage - each letter of the sequence has assigned
an integer (its index in alphabet of \code{sq} object). Those integers in binary format
fit in less than 8 bits, but normally are stored on 16 bits. However, thanks to bit
packing it is possible to remove unused bits and store numbers more tightly. This 
operations result in a little time overhead in all operations, because most of them 
require unpacking and repacking sequences, but this cost is relatively low in comparison
to amount of saved memory.

For example - \strong{dna} \strong{cln} alphabet consist of 6 values: ACGT-. They are 
assigned numbers 1 to 5 respectively. Those numbers in binary format take form: \code{001},
\code{010}, \code{011}, \code{100}, \code{101}. Each of the letters can  be coded with
just 3 bits instead of 8 which is demanded by \code{char} - this allows
us to save more than 60\% of memory spent on storage of nucleotide sequences.
}

\section{tibble compatibility}{

\code{sq} objects are compatible with \code{\link[tibble]{tibble}} class - that means one can 
have a \code{sq} object as a column of a \code{tibble}. There are overloaded print methods, so
that it is printed in pretty format.
}

