% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqapply.R
\name{sqapply}
\alias{sqapply}
\title{Apply function to each sequence}
\usage{
sqapply(
  x,
  fun,
  ...,
  single_string = FALSE,
  NA_letter = getOption("tidysq_NA_letter")
)
}
\arguments{
\item{x}{[\code{sq}]\cr
An object this function is applied to.}

\item{fun}{[\code{function(1)}]\cr
A function to apply to each sequence in \code{sq} object; it should
take a character vector, numeric vector or single character string as an input.}

\item{...}{further arguments to be passed from or to other methods.}

\item{paste_char}{[\code{logical(1)}]\cr
A value indicating in which form sequences should be
passed to the function \code{fun}; if \code{FALSE} (default), they will be treated as character
vectors, if \code{TRUE}, they will be pasted into a single string.}

\item{use_na_letter}{[\code{logical(1)}]\cr
A value indicating whether to use a printing character
to represent \code{\link{NA}} values; if \code{TRUE}, letter from option "tidysq_p_na_letter"
will be used instead of \code{NA} values (default value for this option is "!", for details
see \code{\link{tidysq-options}}), otherwise just \code{NA} values will be used; default value
for this parameter is equal to \code{paste_char} value; \code{use_na_letter} cannot be
\code{FALSE} if \code{paste_char} is \code{TRUE}.}
}
\value{
A list of values returned by function for each sequence in corresponding order.
}
\description{
Applies given function to each sequence. Sequences are passed to function as character vectors
(or numeric, if type of \code{sq} is \strong{enc}) or single character strings, depending on 
parameter.
}
\seealso{
\code{\link[=sq-class]{sq}} \code{\link[base]{lapply}}
}
