% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-methods.R
\name{sqextract}
\alias{sqextract}
\alias{sq-extract}
\title{Extract parts of a sq object}
\arguments{
\item{x}{sq object from which to extract element(s)}

\item{i, j, ...}{indices specifying elements to extract. They may be 
\code{\link{numeric}}, \code{\link{character}} or \code{\link{logical}} vectors or empty. 
This function follows \code{\link[vctrs]{vctrs-package}} conventions regarding argument
interpretation for indexing vectors, which are a bit stricter that normal R
conventions, for example implicit argument recycling is prohibited.}
}
\value{
\code{\link[=sq-class]{sq}} object of the same type as input sq, containing
extracted elements
}
\description{
Operator to extract subsets of sq objects.
}
\details{
This function allows extracting specified sequences from the 
sq object and follows the normal R conventions. For details refer to the 
R documentation (see 
\url{https://cran.r-project.org/doc/manuals/r-release/R-intro.html#Index-vectors}). 
Subsetting of the sq object does not affect its attributes (class and alphabet 
of the object). Attempt to extract elements using indices not present in
the object will return an error.
}
\seealso{
\code{\link[=sq-class]{sq}} \code{\link{bite}}
}
