% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-tidysq.R
\docType{package}
\name{tidysq-package}
\alias{tidysq-package}
\alias{tidysq}
\title{tidysq: tidy analysis of biological sequences}
\description{
The \code{tidysq} package is a toolbox for the analysis of 
biological sequences in a tidy way.
}
\author{
Michal Burdukiewicz, Dominik Rafacz, Mateusz Bąkała, Leon Eyrich Jessen
}
