% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Save sq to fasta file}
\usage{
write_fasta(x, name, file, nchar = 80)
}
\arguments{
\item{x}{a \code{\link{sq}} object.}

\item{name}{a \code{\link{character}} vector of length equal to \code{sq} length.}

\item{file}{a \code{\link{character}} string indicating path to file to write into.}

\item{nchar}{a positive \code{\link{integer}} value informing about maximum number of 
characters to put in each line of file.}
}
\description{
Writes \code{\link{sq}} objects with their names to a fasta file.
}
